Supervised-UnsupervisedLearningSmokingDrinkingData
The project uses the "Smoking and Drinking Dataset with Body Signal" dataset. This dataset is collected from the National Health Insurance Service in Korea. All personal information and sensitive data have been excluded. The purpose of this dataset is to:

Analyze body signals and classify individuals as smokers or drinkers.

The dataset is 29 MB in size and contains 991,346 observations and 24 features. Have applied to supervised and unsupervised machine learning methods.